refactor: 코드 개선하기
