+++ src/components/nav/Nav.tsx
@@ -1,9 +1,7 @@
 import styled from 'styled-components';
 import SearchBtn from '../button/SearchBtn';
 import ThemeBtn from '../button/ThemeBtn';
-import { useDispatch, useSelector } from 'react-redux';
-import { themeActions } from '../../store/theme-slice';
-import { RootState } from '../../store';
+import useToggleTheme from '../../hooks/useToggleTheme';
 
 const NavContainer = styled.nav``;
 
@@ -14,24 +12,7 @@ const MenuList = styled.ul`
 `;
 
 const Nav = () => {
-  const dispatch = useDispatch();
-  const ThemeMode = useSelector((state: RootState) => state.themeType.themeMode);
-
-  const themeSave = (value: 'light' | 'dark') => {
-    localStorage.setItem('THEME', value);
-  };
-
-  const handlerTheme = () => {
-    if (!ThemeMode) return;
-    if (ThemeMode === 'dark') {
-      dispatch(themeActions.toggleLightMode());
-      themeSave('light');
-    } else {
-      dispatch(themeActions.toggleDarkMode());
-      themeSave('dark');
-    }
-  };
-
+  const handlerTheme = useToggleTheme();
   return (
     <NavContainer>
       <MenuList>
